requirements.txt
Flask==2.3.3
supabase==1.1.0
requests==2.32.3
python-vlc==3.0.18122
Pillow==10.3.0
pytz==2023.3
pystray==0.19.5
setuptools>=69.5.1
PySide6==6.6.0
python-dotenv==1.0.1
werkzeug==3.0.4
importlib-metadata==8.5.0
python-dateutil==2.9.0.post0
mysql-connector-python==8.3.0
msal==1.31.0
